index.html 	Добавлена ссылка на страницу загрузки
upload.php
	Множественные визуальные изменения
	Переработана форма отправки

+++ upload.php
@@ -36,31 +36,31 @@ mysql_select_db("asterisk") or die(mysql_error());
 
  if ($_SERVER["REQUEST_METHOD"] == "POST" ) {
  // Инициализация переменных
- $target_dir = "/var/lib/asterisk/sounds";
+ $target_dir = "/var/www/html/sounds";
  $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
  $uploadOk = 1;
  $FileType = pathinfo($target_file,PATHINFO_EXTENSION);
 // Проверка расширения
-	if($FileType != "wav"  ) {
+	if($FileType != "wav" && empty($_POST['delete_code']) && $_FILES["fileToUpload"]["type"] != "audio/x-wav") {
     echo '<div id="warning">Загрузи файл в формате WAV!</div>';
     $uploadOk = 0;
 }
 // Проверка дубликата 
-if (file_exists($target_file)) {
+if (file_exists($target_file) && empty($_POST['delete_code'])) {
     echo '<div id="warning">Такой файл уже существует!</div>';
     $uploadOk = 0;
 } 
 // Проверка размера файла
-if ($_FILES["fileToUpload"]["size"] > 2000000) {
+if ($_FILES["fileToUpload"]["size"] > 2000000 && empty($_POST['delete_code'])) {
     echo '<div id="warning">Файл слишком большой!</div>';
     $uploadOk = 0;
 }
 // Загрузка файла
-if ($uploadOk == 0) {
+if ($uploadOk == 0 && empty($_POST['delete_code'])) {
     echo '<div id="warning">Ошибка. Файл не добавлен!</div>';
-} else {
+} elseif (empty($_POST['delete_code'])) {
     if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
-        echo "Файл ". basename( $_FILES["fileToUpload"]["name"]). " был добавлен.";
+        echo "<h2>Файл ". basename( $_FILES["fileToUpload"]["name"]). " был добавлен.</h2>";
         // Коммит в базу
         mysql_query('insert into alarm_codes values("'.$_FILES["fileToUpload"]["name"].'","'.$_POST['header'].'","'.$_POST['message'].'")') or die(mysql_error());  
     } else {
@@ -71,33 +71,40 @@ if ($uploadOk == 0) {
  //Удаление из базы
  if(isset($_POST['delete_code'])) {
 	 mysql_query("DELETE FROM alarm_codes WHERE alarm_code = '".$_POST['delete_code']."'");
+	 $file_path = $target_dir . $_POST['delete_code'];
+	 if (file_exists($file_path)) {
+        unlink($file_path);
+    }
  }
 }
 
  $sql_data = mysql_query("select * from alarm_codes") or die(mysql_error());
  
   $number = 1;
+  print '<form method="post" action="'.htmlspecialchars($_SERVER["PHP_SELF"]) .'" onsubmit="';
+        print "return confirm('Вы уверены?');";
+                print '">';
 Print "<table border cellpadding=3 style=width:100% algin=center>";
-Print "<th>№</th><th>Файл</th><th>Заголовок сообщения</th><th>Текст сообщения</th><th>Удалить</th>";
+Print "<th>№</th><th>Файл</th><th>Тема сообщения</th><th>Текст сообщения</th><th>Удалить</th>";
  while($alarm = mysql_fetch_array( $sql_data )) 
  { 
-	 $name=preg_split("/[\.wav,]+/", $alarm['alarm_code']);
+	 
 	  Print "<tr>";
 	   Print "<td>".$number."</td> ";
-	   Print "<td>".$name[0] . " </td> ";
+	   Print "<td>".$alarm['alarm_code'] . " </td> ";
 	    Print "<td>".$alarm['header']."</td> ";
 	    Print "<td>".$alarm['message']."</td> ";
 	     Print '<td><input type="checkbox" name="delete_code" value="'.$alarm['alarm_code'].'" /></td></tr>'; 
          $number++;
  }
- Print "</table>"; 
+ Print '</th></table><input type="submit" name="submit" value="Удалить"></form>'; 
 mysql_close($mysql);
  ?><hr>
- <div align=center>	<h2>Загрузка файлов в систему автоматического оповещения</h2></div>
+ <div align=center>	<h2>Загрузка файлов в систему автоматического оповещения</h2></div><br>
 <form action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post" enctype="multipart/form-data">
     <h2>Выбери файл для загрузки:</h2>
     <input type="file" name="fileToUpload" id="fileToUpload">    
-<h2>Параметры файла: 8кгц моно wav</h2>
+<h3>Параметры файла: 8кгц моно wav</h3><hr><br>
   <h2>Настройки текстового оповещения</h2>
 <hr>
 	<h2>Введи тему сообщения:</h2>
@@ -105,7 +112,12 @@ mysql_close($mysql);
 <h2>Введи текст сообщения:</h2>
 <textarea name="message" rows="5" cols="64"></textarea>
 
-<input type="submit" value="Upload" name="Отправить">
+ <table border=0 cellpadding=4 algin=center > 
+  <TR> 
+	  <TD BGCOLOR="#FF0000"><input type="submit" value="Отправить" name="Upload"></TD>
+	  <TD BGCOLOR="#FF0000"><input type="reset" value="СБРОС"></TD>
+  </TR>
+  </table>
 </form>
   <hr>
